Moving censor code to it’s own module
